Nest the `/session-cookie` endpoint under `/api`
